Add DSP_LIBRARY_PATH mappings for target machines
Add DSP_LIBRARY_PATH entries in a new YAML file for the
supported machines to enable consistent, machine-specific
DSP library resolution.

The machine entries are sorted alphabetically by machine name
to ensure maintainability and stability as new properties are
added in the future.

Update the Makefile to install the YAML file to the conf.d
directory alongside the DSP binaries.

Signed-off-by: Vinayak Katoch <vkatoch@qti.qualcomm.com>